Type Definition ppv::f64x4

type f64x4 = Simd<[f64; 4]>;

A 256-bit vector with 4 f64 lanes.

Methods

impl f64x4

pub const fn new(x0: f64, x1: f64, x2: f64, x3: f64) -> Self

Creates a new instance with each vector elements initialized with the provided values.

pub const fn lanes() -> usize

Returns the number of vector lanes.

pub const fn splat(value: f64) -> Self

Constructs a new instance with each element initialized to value.

pub fn extract(self, index: usize) -> f64

Extracts the value at index.

Panics

If index >= Self::lanes().

pub unsafe fn extract_unchecked(self, index: usize) -> f64

Extracts the value at index.

Precondition

If index >= Self::lanes() the behavior is undefined.

#[must_use = "replace does not modify the original value - it returns a new vector with the value at `index` replaced by `new_value`d"]
pub fn replace(self, index: usize, new_value: f64) -> Self

Returns a new vector where the value at index is replaced by new_value.

Panics

If index >= Self::lanes().

#[must_use = "replace_unchecked does not modify the original value - it returns a new vector with the value at `index` replaced by `new_value`d"]
pub unsafe fn replace_unchecked(self, index: usize, new_value: f64) -> Self

Returns a new vector where the value at index is replaced by new_value.

Precondition

If index >= Self::lanes() the behavior is undefined.

impl f64x4

pub fn min(self, x: Self) -> Self

Minimum of two vectors.

Returns a new vector containing the minimum value of each of the input vector lanes.

pub fn max(self, x: Self) -> Self

Maximum of two vectors.

Returns a new vector containing the maximum value of each of the input vector lanes.

impl f64x4

pub fn sum(self) -> f64

Horizontal sum of the vector elements.

The intrinsic performs a tree-reduction of the vector elements. That is, for an 8 element vector:

((x0 + x1) + (x2 + x3)) + ((x4 + x5) + (x6 + x7))

If one of the vector element is NaN the reduction returns NaN. The resulting NaN is not required to be equal to any of the NaNs in the vector.

pub fn product(self) -> f64

Horizontal product of the vector elements.

The intrinsic performs a tree-reduction of the vector elements. That is, for an 8 element vector:

((x0 * x1) * (x2 * x3)) * ((x4 * x5) * (x6 * x7))

If one of the vector element is NaN the reduction returns NaN. The resulting NaN is not required to be equal to any of the NaNs in the vector.

impl f64x4

pub fn max_element(self) -> f64

Largest vector element value.

pub fn min_element(self) -> f64

Smallest vector element value.

impl f64x4

pub fn from_slice_aligned(slice: &[f64]) -> Self

Instantiates a new vector with the values of the slice.

Panics

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary.

pub fn from_slice_unaligned(slice: &[f64]) -> Self

Instantiates a new vector with the values of the slice.

Panics

If slice.len() < Self::lanes().

pub unsafe fn from_slice_aligned_unchecked(slice: &[f64]) -> Self

Instantiates a new vector with the values of the slice.

Precondition

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary, the behavior is undefined.

pub unsafe fn from_slice_unaligned_unchecked(slice: &[f64]) -> Self

Instantiates a new vector with the values of the slice.

Precondition

If slice.len() < Self::lanes() the behavior is undefined.

impl f64x4

pub fn write_to_slice_aligned(self, slice: &mut [f64])

Writes the values of the vector to the slice.

Panics

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary.

pub fn write_to_slice_unaligned(self, slice: &mut [f64])

Writes the values of the vector to the slice.

Panics

If slice.len() < Self::lanes().

pub unsafe fn write_to_slice_aligned_unchecked(self, slice: &mut [f64])

Writes the values of the vector to the slice.

Precondition

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary, the behavior is undefined.

pub unsafe fn write_to_slice_unaligned_unchecked(self, slice: &mut [f64])

Writes the values of the vector to the slice.

Precondition

If slice.len() < Self::lanes() the behavior is undefined.

impl f64x4

pub fn abs(self) -> Self

Absolute value.

impl f64x4

pub fn cos(self) -> Self

Cosine.

impl f64x4

pub fn fma(self, y: Self, z: Self) -> Self

Fused multiply add: self * y + z

impl f64x4

pub fn rsqrte(self) -> Self

impl f64x4

pub fn sin(self) -> Self

impl f64x4

pub fn sqrt(self) -> Self

impl f64x4

pub fn sqrte(self) -> Self

impl f64x4

pub fn eq(self, other: f64x4) -> m64x4

Lane-wise equality comparison.

pub fn ne(self, other: f64x4) -> m64x4

Lane-wise inequality comparison.

pub fn lt(self, other: f64x4) -> m64x4

Lane-wise less-than comparison.

pub fn le(self, other: f64x4) -> m64x4

Lane-wise less-than-or-equals comparison.

pub fn gt(self, other: f64x4) -> m64x4

Lane-wise greater-than comparison.

pub fn ge(self, other: f64x4) -> m64x4

Lane-wise greater-than-or-equals comparison.